requirements.txt
# Core
streamlit==1.38.0
python-dotenv==1.0.1
pydantic==2.8.2
Pillow==9.5.0
bcrypt==4.0.1

# Google Login (Authlib + Flask, replacing streamlit-oauth)
Flask==3.0.3
authlib==1.3.0
requests==2.32.3

# LLM / AI
google-generativeai==0.8.2

requests==2.32.3
requests-oauthlib==2.0.0
